Section 4 endpoints; unit tests for api; added utils.exceptions

+++ app/api/v1/tasks.py
@@ -1,6 +1,6 @@
-from fastapi import HTTPException, Depends, APIRouter
+from fastapi import Depends, APIRouter, status
 from sqlalchemy.ext.asyncio import AsyncSession
-from app.schemas.tasks import TaskResponseList, TaskResponse, TaskBase
+from app.schemas.tasks import TaskResponseList, TaskResponse, TaskBase, TaskPartialUpdate
 from app.db.models.users import UserORM
 from app.db.database import get_db
 from app.services.auth import get_current_user
@@ -12,7 +12,11 @@ from app.services.tasks import (
     save_task_in_db,
     update_task_obj,
     delete_task_from_db,
+    filter_tasks_by_status,
+    check_task_owner,
 )
+from app.utils.enums import TaskStatus
+from app.utils.exceptions import RecordNotFound
 
 
 router = APIRouter(prefix="/tasks")
@@ -54,7 +58,7 @@ async def list_user_tasks(
     return tasks_list
 
 
-@router.post("", response_model=TaskResponse)
+@router.post("", response_model=TaskResponse, status_code=status.HTTP_201_CREATED)
 async def create_task(
     task_data: TaskBase,
     db: AsyncSession = Depends(get_db),
@@ -73,7 +77,7 @@ async def create_task(
 @router.patch("/{task_id}", response_model=TaskResponse)
 async def partial_update_task(
     task_id: int,
-    task_data: TaskBase,
+    task_data: TaskPartialUpdate,
     db: AsyncSession = Depends(get_db),
     user_obj: UserORM = Depends(get_current_user),
 ):
@@ -81,12 +85,15 @@ async def partial_update_task(
     Here only authorized owner of the task can update it **IN WORK**
     """
     task_obj = await get_task_by_id(db=db, task_id=task_id)
-    await update_task_obj(db=db, update_data=task_data.model_dump(), task_obj=task_obj)
-    await db.refresh(task_obj)
-    return task_obj
+    if check_task_owner(db=db, task_obj=task_obj, user_id=user_obj.id):
+        await update_task_obj(db=db, update_data=task_data.model_dump(), task_obj=task_obj)
+        await db.refresh(task_obj)
+        return task_obj
+    else:
+        raise RecordNotFound()
 
 
-@router.delete("/{task_id}")
+@router.delete("/{task_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def destroy_task(
     task_id: int, db: AsyncSession = Depends(get_db), user_obj: UserORM = Depends(get_current_user)
 ):
@@ -94,5 +101,41 @@ async def destroy_task(
     Here only authorized owner of the task can destroy it **IN WORK**
     """
     task_obj = await get_task_by_id(db=db, task_id=task_id)
-    await delete_task_from_db(db=db, task_obj=task_obj)
-    return {"status": "success"}
+    if check_task_owner(db=db, task_obj=task_obj, user_id=user_obj.id):
+        await delete_task_from_db(db=db, task_obj=task_obj)
+    else:
+        raise RecordNotFound()
+
+
+@router.post("/{task_id}", response_model=TaskResponse)
+async def mark_task_as_completed(
+    task_id: int, db: AsyncSession = Depends(get_db), user_obj: UserORM = Depends(get_current_user)
+):
+    """
+    Mark given task as completed. Section 4 of tech task
+    """
+    task_obj = await get_task_by_id(db=db, task_id=task_id)
+    if check_task_owner(db=db, task_obj=task_obj, user_id=user_obj.id):
+        await update_task_obj(
+            db=db, update_data={"status": TaskStatus.COMPLETED}, task_obj=task_obj
+        )
+        await db.refresh(task_obj)
+        return task_obj
+    else:
+        raise RecordNotFound()
+
+
+@router.get("/filter/", response_model=TaskResponseList)
+async def get_tasks_filtered_by_status(
+    task_status: TaskStatus = None,
+    db: AsyncSession = Depends(get_db),
+    user_obj: UserORM = Depends(get_current_user),
+):
+    """
+    Filter tasks by status. Section 4 of teck task
+    """
+    if task_status:
+        tasks_list = await filter_tasks_by_status(db=db, task_status=task_status)
+    else:
+        tasks_list = await get_all_tasks(db=db)
+    return tasks_list
